requirements.txt



flask
requests

Werkzeug
google-generativeai
PyGithub
python-dotenv
pdfplumber

